move files around and nowrap
