+++ app/views/pages/home.html.erb
@@ -18,4 +18,4 @@
 
 <div id="root"></div>
 <div id="root2"></div>
-<div id="root3"></div>
+<div id="Album"></div>
